Password.jsx renamed to PasswordInput and component name fixed in tests
